+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+0.12.2
+
+- fixed xeditable double values
+- fixed xls export for consortia members
+- fixed and improved finance xls export
+
 0.12.1
 
 - temporary deactivated usage and statistics